requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
psycopg[binary]==3.2.3
numpy==2.1.3
gunicorn==22.0.0

# sentiment modülün dış kütüphane kullanıyorsa ekle (ör: transformers/torch vb.)
sentence-transformers==3.0.1
torch==2.4.1
transformers==4.44.2
